Allow `tau` to be parameterized in Search
This adds a new SearchParameters struct which is passed in to the Search
function. Rather than just a max results int, we now also allow a
MaxDistance field, which is assigned to `tau` in the Search function.
This allows you to only consider items that are closer than a certain
maximum distance threshold.